move Direction class to Oneway.kt
