+++ app/src/main/java/de/westnordost/streetcomplete/osm/Oneway.kt
@@ -1,6 +1,6 @@
 package de.westnordost.streetcomplete.osm
 
-import de.westnordost.streetcomplete.osm.cycleway.Direction
+import kotlinx.serialization.Serializable
 
 fun isForwardOneway(tags: Map<String, String>): Boolean =
     tags["oneway"] == "yes"
@@ -25,3 +25,22 @@ fun isNotOnewayForCyclists(tags: Map<String, String>, isLeftHandTraffic: Boolean
 fun isInContraflowOfOneway(tags: Map<String, String>, direction: Direction): Boolean =
     isForwardOneway(tags) && direction == Direction.BACKWARD
     || isReversedOneway(tags) && direction == Direction.FORWARD
+
+@Serializable
+enum class Direction {
+    FORWARD,
+    BACKWARD,
+    BOTH;
+
+    fun reverse(): Direction = when (this) {
+        FORWARD -> BACKWARD
+        BACKWARD -> FORWARD
+        BOTH -> BOTH
+    }
+
+    companion object {
+        /** Return the default direction of a oneway if nothing is specified */
+        fun getDefault(isRightSide: Boolean, isLeftHandTraffic: Boolean): Direction =
+            if (isRightSide xor isLeftHandTraffic) FORWARD else BACKWARD
+    }
+}
